escola/requirements.txt
psycopg2-binary==2.9.9
Django
djangorestframework
django-cors-headers